requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
fpdf2
python-multipart
python-dotenv